Alex/update to fix new xcode issues (#150)
Multiple issues were found when attempting to update the pod trunk after merging in PR #147.

These issues were caused by a few things:
- flatMap needed to switch to compactMap in three places (ref: https://github.com/apple/swift-evolution/blob/master/proposals/0187-introduce-filtermap.md)
- The podspec still showed support for iOS 8.0 which prevented the use of NSLayoutAnchors (pod spec linting failed)
- Our project file still showed support for Swift 2.3 in the main proj file, while the individual targets were flagged for 4.1 already.

The 0.9.2 tag has not been released to Cocoapods, but exists in the repo. This tag currently points to a broken version should anyone manually install the project and use Xcode 9.3 (Swift 4.1). After the PR is merged should the tag be rewritten to point to this commit (overwriting 0.9.2), or should that tag stay around?